requirements.txt
# Core Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database
chromadb>=0.4.22

# ML/AI
# torch: install a CUDA build matching your GPU/driver (see docs/setup/gpu-setup.md).
# The CPU wheel is fine for fake-models tests but SAM3/Florence-2 real inference wants GPU.
torch>=2.1.0
# transformers: Florence-2 custom code is written against the 4.x API.
# Upper-bounding until Florence-2 upstream supports 5.x.
transformers>=4.36.0,<5.0
# Florence-2 trust_remote_code loader depends on these at import time
einops>=0.8.0
timm>=1.0.0
# SAM3 (vendor/sam3) depends on these
iopath>=0.1.10
ftfy>=6.1.1
pycocotools>=2.0.7
# Note: SAM3 perflib also imports `triton`. triton is platform-dependent:
#   Linux + CUDA:   pip install triton
#   Windows + CUDA: pip install triton-windows
# Not listed here so the same file works on CPU-only CI.
# See docs/setup/gpu-setup.md for details.
ultralytics>=8.1.0
opencv-python-headless>=4.9.0
pillow>=10.2.0
numpy>=1.26.0

# Data Processing
pandas>=2.1.0

# Experiment Tracking
mlflow>=2.10.0

# UI
streamlit>=1.31.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
python-dotenv>=1.0.0
